+++ public/src/components/ProfilePage/FilmTab/style.css
@@ -1,0 +1,136 @@
+/* LARGE FILMS ==============================*/
+.large-film-container {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  column-gap: 20px;
+  width: 300px;
+  overflow-x: auto;
+  white-space: nowrap;
+}
+.large-film-tile {
+  position: relative;
+}
+.large-film-img,
+.empty-tile-btn {
+  width: 131px;
+  height: 178px;
+}
+
+.empty-tile-btn {
+  background: #ebebeb;
+  outline: none;
+  border: 1px dashed #a7a6a6;
+  border-radius: 4px;
+  display: inline-block;
+}
+/* SMALL FILMS ==============================*/
+.small-tile-container {
+  margin-top: 2rem;
+  display: flex;
+  flex-wrap: wrap;
+  width: 300px;
+}
+
+.small-film-img {
+  width: 100px;
+}
+.small-tile-section {
+  margin-top: 2rem;
+}
+.cropped-to-square {
+  overflow: hidden;
+  width: 100px;
+  height: 100px;
+  margin-right: 15px;
+  margin-bottom: 15px;
+  position: relative;
+}
+
+.tag-group {
+  width: 100%;
+}
+
+.tag-group-film-container {
+  display: flex;
+  width: 400px;
+  flex-wrap: wrap;
+}
+
+.tag-dropdown {
+  background: white;
+  box-shadow: 0px 3px 15px rgba(0, 0, 0, 0.2);
+  border-radius: 4px;
+  position: absolute;
+  top: 47px;
+  width: 120px;
+  z-index: 2;
+  display: flex;
+  flex-direction: column;
+}
+/* ADD FILM PANEL ==============================*/
+
+.add-film-drop-down {
+  background: white;
+  padding: 2rem;
+  box-shadow: 0px 3px 15px rgba(0, 0, 0, 0.2);
+  border-radius: 4px;
+  position: absolute;
+  top: 20px;
+  z-index: 2;
+  width: 100%;
+  box-sizing: border-box;
+}
+
+.open-add-film-dropdown-btn {
+  position: relative;
+  width: 400px;
+}
+
+.color-square {
+  width: 20px;
+  height: 20px;
+  border: 2px solid lightgrey;
+
+  margin: 10px 0 20px 1rem;
+}
+.color-picker {
+  position: absolute;
+  z-index: 3;
+  left: 22px;
+}
+
+.tag-dropdown-btn {
+  padding: 0.5rem;
+}
+.tag-dropdown-color {
+  border-radius: 4px;
+}
+.color-picker-container {
+  position: relative;
+}
+.tag-input-container {
+  display: flex;
+  margin-bottom: 2rem;
+  position: relative;
+}
+/* GENERAL ============================== */
+.onhover-tile-btn-container {
+  position: absolute;
+  top: 20px;
+  right: 0px;
+  display: flex;
+  flex-direction: column;
+  background: white;
+  opacity: 0.9;
+  width: 95px;
+}
+
+.more-menu-btn {
+  position: absolute;
+  top: 0px;
+  right: 0px;
+  background: white;
+  opacity: 0.9;
+  padding: 0 4px 4px 4px;
+  border-radius: 2px;
+}
