+++ public/src/components/ProfilePage/style.css
@@ -5,10 +5,7 @@ h2 {
   margin-bottom: 20px;
 }
 
-.link-container {
-  display: flex;
-  flex-direction: column;
-}
+/* MODAL=================================== */
 
 .modal-window {
   width: 450px;
@@ -37,88 +34,42 @@ h2 {
   height: 400px;
   display: flex;
   flex-direction: column;
-
   margin: 2rem;
 }
 
-.drop-down-container {
+.modal-btn {
+  height: 35px;
+  width: 49%;
+}
+/* SEND REX MODAL  ======================*/
+.modal-btn-container {
+  display: flex;
+  justify-content: space-between;
+}
+
+/* FILM DROP DOWN ============================ */
+.film-drop-down-container {
   background: white;
   display: flex;
   flex-direction: column;
   box-shadow: 0px 3px 15px rgba(0, 0, 0, 0.2);
   position: absolute;
   z-index: 2;
-}
-
-.add-film-drop-down {
-  background: white;
-  padding: 2rem;
-  box-shadow: 0px 3px 15px rgba(0, 0, 0, 0.2);
-  border-radius: 4px;
-  position: absolute;
-  top: 20px;
-  z-index: 2;
-  width: 100%;
-  box-sizing: border-box;
-}
-.tag-group-film-container {
-  display: flex;
-  width: 400px;
-  flex-wrap: wrap;
-}
-.tag-dropdown {
-  background: white;
-  box-shadow: 0px 3px 15px rgba(0, 0, 0, 0.2);
-  border-radius: 4px;
-  position: absolute;
-  top: 47px;
-  width: 120px;
-  z-index: 2;
-  display: flex;
-  flex-direction: column;
-}
-.open-add-film-dropdown-btn {
-  position: relative;
-
-  width: 400px;
-}
-.color-square {
-  width: 20px;
-  height: 20px;
-  border: 2px solid lightgrey;
-
-  margin: 10px 0 20px 1rem;
-}
-.color-picker {
-  position: absolute;
-  z-index: 3;
-  left: 22px;
-}
-
-.tag-dropdown-btn {
-  padding: 0.5rem;
-}
-.tag-dropdown-color {
-  border-radius: 4px;
-}
-.color-picker-container {
-  position: relative;
-}
-.tag-input-container {
-  display: flex;
-  margin-bottom: 2rem;
-  position: relative;
-}
-.dropdown-list-item {
-  display: flex;
-}
-.auto-complete-list {
-  list-style: none;
 }
 
 .drop-down-image {
   width: 42px;
 }
+
+.auto-complete-list {
+  list-style: none;
+}
+
+.dropdown-list-item {
+  display: flex;
+}
+
+/* PROFILE GENERIC =============================*/
 
 .profile-pic {
   border: 2px solid lightgrey;
@@ -129,7 +80,6 @@ h2 {
   display: flex;
   margin-bottom: 2rem;
 }
-
 .send-rex-btn {
   height: 20px;
 }
@@ -164,101 +114,7 @@ h2 {
   padding-bottom: 8px;
 }
 
-.rex-list-item-info {
-  margin-left: 1rem;
-  padding: 1rem 0;
-}
-
-.rex-list-item {
-  display: flex;
-  margin-bottom: 2rem;
-  background: #ebebeb;
-}
-.rex-info-p-tag {
-  margin: 0.5rem 0;
-  font-size: 0.7rem;
-}
-
-.rex-list {
-  margin-top: 3rem;
-}
-
-.rex-img {
-  width: 70px;
-}
-
-.large-film-container {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  column-gap: 20px;
-  width: 300px;
-  overflow-x: auto;
-  white-space: nowrap;
-}
-.small-tile-container {
-  margin-top: 2rem;
-  display: flex;
-  flex-wrap: wrap;
-  width: 300px;
-}
-.large-film-img,
-.empty-tile-btn {
-  width: 131px;
-  height: 178px;
-}
-
-.empty-tile-btn {
-  background: #ebebeb;
-  outline: none;
-  border: 1px dashed #a7a6a6;
-  border-radius: 4px;
-  display: inline-block;
-}
-.small-film-img {
-  width: 100px;
-}
-.small-tile-section {
-  margin-top: 2rem;
-}
-
-.cropped-to-square {
-  overflow: hidden;
-  width: 100px;
-  height: 100px;
-  margin-right: 15px;
-  margin-bottom: 15px;
-  position: relative;
-}
-
-.onhover-tile-btn-container {
-  position: absolute;
-  top: 20px;
-  right: 0px;
-  display: flex;
-  flex-direction: column;
-  background: white;
-  opacity: 0.9;
-  width: 95px;
-}
-
-.large-film-tile {
-  position: relative;
-}
-
-.more-menu-btn {
-  position: absolute;
-  top: 0px;
-  right: 0px;
-  background: white;
-  opacity: 0.9;
-  padding: 0 4px 4px 4px;
-  border-radius: 2px;
-}
-
-.modal-btn-container {
-  display: flex;
-  justify-content: space-between;
-}
+/* GENERAL LABELS INPUTS BUTTONS ============== */
 
 .send-rex-label,
 .film-dropdown-label,
@@ -270,20 +126,6 @@ h2 {
   height: 35px;
   width: 100%;
   margin: 10px 0 20px 0;
-}
-
-.tag-input {
-  margin: 10px 0 20px 0;
-  height: 35px;
-  width: 40%;
-}
-.modal-btn::placeholder {
-  font-family: 'Red Hat Text', sans-serif;
-}
-
-.modal-btn {
-  height: 35px;
-  width: 49%;
 }
 
 .button {
@@ -301,12 +143,9 @@ h2 {
 
 .text-input::placeholder {
   padding-left: 5px;
+  font-family: 'Red Hat Text', sans-serif;
 }
 
 .interactive {
   cursor: pointer;
-}
-
-.tag-group {
-  width: 100%;
 }
